fixed returns on retries to not immediatly complete resolves #549
