+++ lib/readers/elasticsearch_date_range/slicer.js
@@ -8,14 +8,16 @@ var dateOptions = require('./../../utils/date_utils').dateOptions;
 var dateFormatMS = require('./../../utils/date_utils').dateFormat;
 var dateFormatS = require('./../../utils/date_utils').dateFormatSeconds;
 var parseError = require('../../utils/error_utils').parseError;
+var retryModule = require('../../utils/error_utils').retryModule;
 
 function newSlicer(context, opConfig, job, retryData, logger, client) {
     var events = context.foundation.getEventEmitter();
     var jobConfig = job.jobConfig;
     var isPersistent = jobConfig.lifecycle === 'persistent';
     var slicers = [];
-    var numOfRetries = job.max_retries;
     var time_resolution = dateOptions(opConfig.time_resolution);
+    var retryError = retryModule(logger, job.max_retries);
+
 
     var dateFormat = time_resolution === 'ms' ? dateFormatMS : dateFormatS;
 
@@ -293,26 +295,6 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
         return getIdData(idSubslicer)
     }
 
-    function retryError(retry, dateObj, err, fn, msg) {
-        var errMessage = parseError(err);
-        logger.error('error while getting next slice', errMessage);
-        var startKey = dateObj.start.format(dateFormat);
-
-        if (!retry[startKey]) {
-            retry[startKey] = 1;
-            fn(msg)
-        }
-        else {
-            retry[startKey] += 1;
-            if (retry[startKey] > numOfRetries) {
-                return Promise.reject(`max_retries met for slice, start: ${startKey}`, errMessage);
-            }
-            else {
-                fn(msg)
-            }
-        }
-    }
-
     function nextChunk(opConfig, client, jobConfig, dates, slicer_id, retryData) {
         var shouldDivideByID = opConfig.subslice_by_key;
         var threshold = opConfig.subslice_key_threshold;
@@ -329,7 +311,6 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
         dateParams.end = moment(dateParams.start.format(dateFormat)).add(dateParams.interval[0], dateParams.interval[1]);
         logger.debug('all date configurations for date slicer', dateParams);
         //used to keep track of retried queries
-        var retry = {};
 
         return function sliceDate(msg) {
             if (dateParams.start.isSameOrAfter(dateParams.limit)) {
@@ -368,7 +349,7 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
                         }
                     })
                     .catch(function(err) {
-                        return retryError(retry, dateParams, err, sliceDate, msg)
+                        return retryError(dateParams.start.format(dateFormat), err, sliceDate, msg)
                     })
             }
         };
@@ -436,9 +417,6 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
 
         logger.debug('all date configurations for date slicer', dateParams);
 
-        //used to keep track of retried queries
-        var retry = {};
-
         //set a timer to add the next set it should process
         setInterval(function() {
             //keep a list of next batches in cases current batch is still running
@@ -495,7 +473,7 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
                             }
                         }
                             .catch(function(err) {
-                                return retryError(retry, dateParams, err, sliceDate, msg)
+                                return retryError(dateParams.start.format(dateFormat), err, sliceDate, msg)
                             })
                     );
             }
