+++ lib/readers/id_slicer.js
@@ -2,6 +2,7 @@
 
 var _ = require('lodash');
 var parseError = require('../utils/error_utils').parseError;
+var retryModule = require('../utils/error_utils').retryModule;
 
 var base64url = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w',
     'x', 'y', 'z', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X',
@@ -18,27 +19,7 @@ module.exports = function(client, job, opConfig, logger, retryData, range) {
     var baseKeyArray = getKeyArray(opConfig);
     var keyArray = opConfig.key_range ? opConfig.key_range : baseKeyArray.slice();
     var elasticsearch = require('elasticsearch_api')(client, logger, opConfig);
-    var numOfRetries = job.max_retries;
-    var retry = {};
-
-    function retryKey(key, err, fn, msg) {
-        var errMessage = parseError(err);
-        logger.error('error while getting next slice', errMessage);
-
-        if (!retry[key]) {
-            retry[key] = 1;
-            fn(msg)
-        }
-        else {
-            retry[key] += 1;
-            if (retry[key] > numOfRetries) {
-                return Promise.reject(`max_retries met for slice, key: ${key}`, errMessage);
-            }
-            else {
-                fn(msg)
-            }
-        }
-    }
+    var retryError = retryModule(logger, job.max_retries);
 
     function getCountForKey(query) {
         return elasticsearch.search(query);
@@ -106,7 +87,7 @@ module.exports = function(client, job, opConfig, logger, retryData, range) {
 
                 })
                 .catch(function(err) {
-                    return retryKey(key, err, getKeySlice, query)
+                    return retryError(key, err, getKeySlice, query)
                 })
         }
