+++ lib/utils/error_utils.js
@@ -17,6 +17,29 @@ function parseError(err) {
     return err.response ? err.response : err;
 }
 
+function retryModule(logger, numOfRetries) {
+    let retry = {};
+    return function(key, err, fn, msg) {
+        var errMessage = parseError(err);
+        logger.error('error while getting next slice', errMessage);
+
+        if (!retry[key]) {
+            retry[key] = 1;
+            return fn(msg)
+        }
+        else {
+            retry[key] += 1;
+            if (retry[key] > numOfRetries) {
+                return Promise.reject(`max_retries met for slice, key: ${key}`, errMessage);
+            }
+            else {
+                return fn(msg)
+            }
+        }
+    }
+}
+
 module.exports = {
-    parseError: parseError
+    parseError: parseError,
+    retryModule: retryModule
 };
